Created simple rectangle and grid detection
Looks through the image and robustly detects the 2 largest contours, and then draws grid lines within them to find the cells. Right now, a current bug is that there are thick sidebars, in which cells are detected within them too. Planning to add horizontal and vertical kernel detection to detect intersections instead of individual cells.